Rename `ImageScaleService` to `ImageScalingService`
