+++ pillarbox-core-business/src/main/java/ch/srgssr/pillarbox/core/business/images/ImageScalingService.kt
@@ -9,7 +9,7 @@ import androidx.annotation.Px
 /**
  * Service used to get a scaled image URL.
  */
-interface ImageScaleService {
+interface ImageScalingService {
     /**
      * The supported widths.
      *
