+++ pillarbox-core-business/src/test/java/ch/srgssr/pillarbox/core/business/MediaCompositionMediaItemSourceTest.kt
@@ -8,7 +8,7 @@ import androidx.media3.common.MediaItem
 import androidx.media3.common.MediaMetadata
 import ch.srgssr.pillarbox.core.business.exception.BlockReasonException
 import ch.srgssr.pillarbox.core.business.exception.ResourceNotFoundException
-import ch.srgssr.pillarbox.core.business.images.DefaultImageScaleService
+import ch.srgssr.pillarbox.core.business.images.DefaultImageScalingService
 import ch.srgssr.pillarbox.core.business.integrationlayer.data.BlockReason
 import ch.srgssr.pillarbox.core.business.integrationlayer.data.Chapter
 import ch.srgssr.pillarbox.core.business.integrationlayer.data.MediaComposition
@@ -23,7 +23,7 @@ class MediaCompositionMediaItemSourceTest {
 
     private val mediaItemSource = MediaCompositionMediaItemSource(
         mediaCompositionDataSource = DummyMediaCompositionProvider(),
-        imageScaleService = DefaultImageScaleService()
+        imageScalingService = DefaultImageScalingService()
     )
 
     @Test(expected = IllegalArgumentException::class)
